Add task to print script output
By default the script output is printed to stdout only if ansible.cfg
has the "stdout_callback = minimal" line or if the playbook is run with
the environment variable "ANSIBLE_STDOUT_CALLBACK=minimal" defined.
This commit adds an explicit printing of the output of the script, so
that the playbook calling this role can be run without the need of
changing ansible.cfg or defining environment variables.