Improve start time logic.
In some cases the stat command doesn't give accurate information on when
the process was started.

Using the ps format we can get the launch time(lstart) of the process.
lstart= ( disables title output)

This fix means that some of the old vms that ceph reported but the
previous version of the tool didn't weren't captured.